drm: Convert settings into CogDrmPlatform properties
Turn the settings for atomic mode setting and the device scale factor
into properties of the CogDrmPlatform class. This allows using
cog_apply_properties_from_key_file() to automatically apply the
settings from the configuration file.

In the future it may be possible to also pass properties when creating
an object with cog_platform_new(), and/or via the "params" argument
of cog_platform_setup(), and always using the same mechanism (object
properties) regardless of where the values come from.